feat(ui): Automatically refresh the deletion challenge to allow 3 attempts
See https://gitlab.com/opensavvy/formulaide/-/merge_requests/26#note_1009838112

+++ ui/src/main/kotlin/formulaide.ui/components/DeletionRequest.kt
@@ -7,14 +7,18 @@ import formulaide.client.routes.deleteRecord
 import formulaide.client.routes.requestDeleteRecord
 import formulaide.ui.components.inputs.Field
 import formulaide.ui.components.inputs.Input
+import formulaide.ui.components.text.ErrorText
 import formulaide.ui.useClient
 import formulaide.ui.utils.classes
+import io.ktor.client.plugins.*
 import react.FC
 import react.Props
 import react.dom.html.InputType
 import react.dom.html.ReactHTML.div
 import react.dom.html.ReactHTML.p
 import react.useState
+
+private const val MAX_ATTEMPTS = 3
 
 internal external interface DeletionRequestProps : Props {
 	var delete: Boolean?
@@ -26,11 +30,13 @@ internal val DeletionRequest = FC<DeletionRequestProps>("DeletionRequest") { pro
 	var challenge by useState<String>()
 	var response by useState<String>()
 
+	var failureCounter by useState(0)
+
 	val (client) = useClient()
 	require(client is Client.Authenticated) { "Les utilisateurs non-connectés ne peuvent pas supprimer des dossiers" }
 	require(client.me.administrator) { "Les non-administrateurs ne peuvent pas supprimer des dossiers" }
 
-	useAsyncEffectOnce {
+	useAsyncEffect(failureCounter) {
 		challenge = client.requestDeleteRecord(props.record.createRef()).challenge
 	}
 
@@ -71,24 +77,45 @@ internal val DeletionRequest = FC<DeletionRequestProps>("DeletionRequest") { pro
 			}
 		}
 
-		StyledButton {
-			text = "Supprimer définitivement"
-			action = {
-				client.deleteRecord(
-					props.record.createRef(),
-					requireNotNull(response) { "Vous n'avez pas rempli la question demandée." }
-				)
+		div {
+			when (failureCounter) {
+				in 0 until MAX_ATTEMPTS -> StyledButton {
+					text = "Supprimer définitivement"
+					action = {
+						try {
+							client.deleteRecord(
+								props.record.createRef(),
+								requireNotNull(response) { "Vous n'avez pas rempli la question demandée." }
+							)
 
-				props.onFinished()
+							props.onFinished()
+						} catch (e: ClientRequestException) {
+							if ("La réponse est incorrecte" in e.message)
+								failureCounter++
+							else
+								throw e
+						}
+					}
+					emphasize = false
+					enabled = challenge != null
+				}
+
+				else -> ErrorText {
+					text = "Trop de tentatives échouées"
+				}
 			}
-			emphasize = false
-			enabled = challenge != null
+
+			StyledButton {
+				text = "Annuler"
+				action = { props.onFinished() }
+				emphasize = true
+			}
 		}
 
-		StyledButton {
-			text = "Annuler"
-			action = { props.onFinished() }
-			emphasize = true
+		if (failureCounter > 0) {
+			ErrorText {
+				text = "Mauvaise réponse. ${MAX_ATTEMPTS - failureCounter} tentative(s) restante(s)."
+			}
 		}
 	}
 }
